+++ Game.Core/Domain/Combat/CombatStateMachine.cs
@@ -1,0 +1,3 @@
+namespace Game.Core.Domain.Combat;
+
+public partial class CombatStateMachine { }