+++ Game.Core/Domain/Combat/States/PlayerTurnState.cs
@@ -1,0 +1,15 @@
+namespace Game.Core.Domain.Combat.States;
+
+public partial class PlayerTurn : ICombatState
+{
+    public CombatStateType Type => CombatStateType.PlayerTurn;
+
+    public void Enter(CombatContext ctx) { }
+
+    public CombatStateTransition Update(CombatContext ctx)
+    {
+        return CombatStateTransition.Next(CombatStateType.PlayerStatus);
+    }
+
+    public void Exit(CombatContext ctx) { }
+}